Generate assert expression repr in runtime
Instead of generating the complete expression repr in compile time, we now
generate code that walks the expression and generates the repr of relevant parts
in runtime, omitting irrelevant parts (the best example is omitting the right-hand
side of an && expression if the left-hand side fails - only the left is shown in
this case).

This change greatly complicates the code, and the emitted runtime code.