Print expression text into "%s", not as the format string itself
It may contain '%' characters. GCC translates printf(...) calls w/o additional arguments
to puts() anyway, but I don't like depending on that. Also it simplies the plugin code a bit
(with the cost of making the runtime code a bit more complex.. Unless GCC do the format in
compile time :)