Fix initialization of 'buf_pos'
Can't get this DECL_INITIAL working. Meh. MODIFY_EXPR works alright.
I attempted to use it on various projects and saw many weird crashes. It ran
as expected when the stack was already initialized, if 'bus_pos' was layed on
on the stack, or when the register used for it was zeroed.