Fix "internal compiler error" on GCC 7.5.0
I had:

    internal compiler error: in gimplify_expr, at gimplify.c:12082

Traced the error to this specific MODIFY_EXPR, guess it was trying to do something with the CALL_EXPR
or one of its descendants.
Wrapping it in a SAVE_EXPR just worked, good enough :shrug: AFAIK it has no harmful effect,
code still does what I want it to do.