Relate names of variables used in assert() to their values
The idea is from how pytest shows it, though I did it a bit different.
Relations are added to the "print list" as the expression tree is walked,
and since it is (now) left-to-right then the print order itself is intuitive.

However since this is done compile-time, it will refer to all variables
in the expression - even those irrelevant to the expression (short circuited).
This won't be okay with functions and more complex subexpressions.