Display repr of function calls with their arguments
Displayed similarly to variables in "subexpressions", but unlike variables which are
aggregated compile-time, functions have to be aggregated only as the expression is
evaluated in run-time (to avoid evaluating them when they're short-circuited).

Thus, their repr is written to a buffer as they are called, and the entire buffer
is printed in the "subexpressions" section.