Register DECL subexpressions for printing only on evaluation
This already exists for functions, because each function subexpression is added
separately to the subexpression buffers (different calls are different subexprs),
so I add them only when they're "used" in the program.

Variables that are used multiple times should be printed once. So previous
code statically collected the list of variables, making sure each appears just once,
and would print them all every time.

New code adds variables to the subexpr buffer when their value is used in a subexpression.
To make sure variables are not printed more than once, each print is wrapped
with a "if was_printed" check, and the variable used is shared with all
prints of the same DECL, so it's printed only if it was used in any subexression
that we're repring, and then it's printed once.